HappyFace/requirements.txt
fastapi~=0.103.0
pydantic~=2.3.0
Pillow~=10.0.0
pymongo~=4.5.0
python-dotenv~=1.0.0
varname~=0.12.0
cryptography~=41.0.4
openai~=0.28.1
uvicorn~=0.23.2